(ref) create separate file for seeding helper functions
