+++ service/src/database/config.js
@@ -7,8 +7,6 @@ const Driver = pgBookshelf.Model.extend({
   location: ['geometry'],
 });
 
-// Driver.prototype.hasTimestamps = true;
-
 const tables = {
   Driver,
 };
