+++ service/src/database/setup/tables.js
@@ -5,8 +5,8 @@ const { pgKnex } = db;
 pgKnex.schema.createTableIfNotExists('drivers', (table) => {
   table.increments();
   table.string('name');
-  table.timestamp('last_checkin');
   table.timestamp('joined');
+  table.timestamp('last_checkin');
   table.boolean('booked');
   table.boolean('available');
   table.string('location');
